web/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
click==8.1.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Security
slowapi==0.1.9

# HTTP Client (for testing)
httpx==0.26.0

# Monitoring
psutil==5.9.8
pynvml==11.5.0

# MedFusion Core
# Note: Install from parent directory
# pip install -e ../../
